+++ tools/api-markdown-documenter/src/index.ts
@@ -38,7 +38,7 @@ export {
 	documentationNodesToHtml,
 	type Transformation as ToHtmlTransformation,
 	type Transformations as ToHtmlTransformations,
-	type TransformationConfig as ToHtmlConfig,
+	type TransformationConfiguration as ToHtmlConfiguration,
 	type TransformationContext as ToHtmlContext,
 } from "./documentation-domain-to-html/index.js";
 export {